requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
transformers>=4.35.0
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
requests==2.31.0
urllib3==2.1.0
youtube-transcript-api==0.6.2
sentencepiece==0.1.99
protobuf==4.25.2
pydantic==1.10.12
accelerate==0.34.2
google-genai
tokenizers>=0.15.0